Add button to increase value
Add a black square as the background UI, the red square is the button to increase the value (later changed), the yellow rectangle is the button for testing (the robotic arm, undecided whether to add it, is the selected button, press once to increase, press the second time to cancel), I originally wanted to add text, but found that it would be covered by the black UI. Then because I was using Chinese records (//), I deleted it in advance. I did not add an English comment and explained it directly here.